fft.go accuracy improved
