+++ user-application/src/components/AudioAnalyser/AudioVisualiser.js
@@ -13,21 +13,20 @@ class AudioVisualiser extends Component {
         const height = canvas1.height;
         const width = canvas1.width;
         const context = canvas1.getContext('2d');
-        const sliceWidth = (width * 2.0)/frequencyData.length;
+        const n = frequencyData.length;
+        const sliceWidth = width/n;
         const upperLimit = 2500;
 
         context.clearRect(0, 0, width, height);
 
-        var n = frequencyData.length;
-
-        const drawRect = (ctx, pos, width, height, colour) => {
+        const drawRect = (ctx, pos, w, h, colour) => {
             ctx.fillStyle = colour;
             ctx.beginPath();
-            ctx.rect(pos, ctx.canvas.height, width, -height);
+            ctx.rect(pos, ctx.canvas.height, w, -h);
             ctx.fill();
           };
 
-        for (var i = 1; i<n/2 + 1; i++) {
+        for (var i = 1; i<n + 1; i++) {
             if (frequencyData[i]>0) {
                 if (frequencyData[i] >= upperLimit) {
                     drawRect(context, i * sliceWidth, sliceWidth, height, 'red');
@@ -45,7 +44,7 @@ class AudioVisualiser extends Component {
       }
 
     render() {
-        return <canvas width="1024" height="500" ref={this.canvas}/>
+        return <canvas width="512" height="500" ref={this.canvas}/>
     }
 }
