+++ user-application/src/components/AudioAnalyser/BackendAudioHandler.js
@@ -1,16 +1,11 @@
 
-async function BackendAudioHandler(dataTimeDomain) {
-
-
-    const audio = await
-        fetch("http://localhost:8080/handleAudio", {
-          method: "POST",
-          body: dataTimeDomain
-        })
-        .then(response => response.json())
-        .then(data => data.list)
-  
-    return audio
+async function BackendAudioHandler(dataTimeDomain) {  
+    return fetch("http://localhost:8080/handleAudio", {
+      method: "POST",
+      body: dataTimeDomain
+    })
+    .then(response => response.json())
+    .then(data => data.list)
 }
 
 export default BackendAudioHandler;
